+++ pyccel/parser/semantic.py
@@ -1356,16 +1356,11 @@ class SemanticParser(BasicParser):
 
             return new_expr
         else:
-            is_inline = func.is_inline if isinstance(func, FunctionDef) else func.functions[0].is_inline
-            if not func.is_semantic:
+            is_inline = func.is_inline if isinstance(func, FunctionDef) else False
+            if is_inline:
+                return self._visit_InlineFunctionDef(func, args, expr)
+            elif not func.is_semantic:
                 func = self._annotate_the_called_function_def(func, args)
-            elif is_inline and isinstance(func, Interface):
-                is_compatible = False
-                for f in func.functions:
-                    fl = self._check_argument_compatibility(args, f.arguments, func, f.is_elemental, raise_error=False)
-                    is_compatible |= fl
-                if not is_compatible:
-                    func = self._annotate_the_called_function_def(func, args)
 
             if self.current_function_name == func.name:
                 if func.results and not isinstance(func.results.var, TypedAstNode):
@@ -1463,6 +1458,7 @@ class SemanticParser(BasicParser):
         func: FunctionDef|Interface
             The new annotated function.
         """
+        assert not old_func.is_inline
         cls_base_syntactic = old_func.get_direct_user_nodes(lambda p: isinstance(p, ClassDef))
         if cls_base_syntactic:
             cls_name = cls_base_syntactic[0].name
@@ -1491,10 +1487,7 @@ class SemanticParser(BasicParser):
 
         # Set the Scope to the FunctionDef's parent Scope and annotate the old_func
         self._scope = new_scope
-        if old_func.is_inline:
-            self._visit_FunctionDef(old_func, function_call_args = function_call_args)
-        else:
-            self._visit_FunctionDef(old_func)
+        self._visit(old_func)
 
         # Retrieve the annotated function
         if cls_base_syntactic:
@@ -2574,7 +2567,14 @@ class SemanticParser(BasicParser):
                 pyccel_stage.set_stage('syntactic')
                 import_node = Import(AsName(mod_name, name))
                 pyccel_stage.set_stage('semantic')
+                # Insert import at global scope
+                current_scope = self.scope
+                scope = current_scope
+                while scope.parent_scope:
+                    scope = scope.parent_scope
+                self.scope = scope
                 self._additional_exprs[-1].append(self._visit(import_node))
+                self.scope = current_scope
                 return self.scope.find(name)
             else:
                 errors.report(f"Unrecognised module {mod_name} imported in global scope. Please import the module locally if it was previously Pyccelised.",
@@ -2974,7 +2974,8 @@ class SemanticParser(BasicParser):
         ls = []
         self._additional_exprs.append([])
         for b in expr.body:
-
+            if isinstance(b, EmptyNode):
+                continue
             # Save parsed code
             line = self._visit(b)
             ls.extend(self._additional_exprs[-1])
@@ -3616,6 +3617,9 @@ class SemanticParser(BasicParser):
             method = d_methods.pop('__init__', None)
 
             if not method.is_semantic:
+                if method.is_inline:
+                    errors.report("An __init__ method cannot be inlined",
+                            severity='fatal', symbol=expr)
                 method = self._annotate_the_called_function_def(method, args)
 
             if method is None:
@@ -4732,7 +4736,12 @@ class SemanticParser(BasicParser):
         else:
             cond        = self._visit(expr.cond)
             value_false = self._visit(expr.value_false)
-            return IfTernaryOperator(cond, value_true, value_false)
+            if isinstance(cond, LiteralTrue):
+                return value_true
+            elif isinstance(cond, LiteralFalse):
+                return value_false
+            else:
+                return IfTernaryOperator(cond, value_true, value_false)
 
     def _visit_Return(self, expr):
 
@@ -4757,7 +4766,10 @@ class SemanticParser(BasicParser):
                         severity='fatal', symbol=expr)
 
         return_objs = func.results
-        return_var = getattr(return_objs.var, 'name', return_objs.var)
+        return_var = return_objs.var
+        if isinstance(return_var, (AnnotatedPyccelSymbol, Variable)):
+            return_var = return_var.name
+
         assigns     = []
         if return_var != results:
             # Create a syntactic object to visit
@@ -4766,9 +4778,12 @@ class SemanticParser(BasicParser):
             pyccel_stage.set_stage('semantic')
 
             a = self._visit(syntactic_assign)
-            assigns.append(a)
-            if isinstance(a, ConstructorCall):
-                a.cls_variable.is_temp = False
+            if not isinstance(a, Assign) or a.lhs != a.rhs:
+                assigns.append(a)
+                if isinstance(a, ConstructorCall):
+                    a.cls_variable.is_temp = False
+            else:
+                a.invalidate_node()
 
         results = self._visit(return_var)
 
@@ -4785,16 +4800,11 @@ class SemanticParser(BasicParser):
             expr  = Return(results)
         return expr
 
-    def _visit_FunctionDef(self, expr, function_call_args=None):
+    def _visit_FunctionDef(self, expr):
         """
-        Annotate the FunctionDef if necessary.
+        Semantically analyse the FunctionDef.
 
-        The FunctionDef is only annotated if the flag annotate is set to True.
-        In the case of an inlined function, we always annotate the function partially,
-        depending on the function call if it is an interface, otherwise we annotate it
-        if the function_call argument are compatible with the FunctionDef arguments.
-        In the case of non inlined function, we only pass through this method
-        twice, the first time we do nothing and the second time we annotate all of functions.
+        Analyse the FunctionDef adding all necessary semantic information.
 
         Parameter
         ---------
@@ -4802,12 +4812,9 @@ class SemanticParser(BasicParser):
            The node that needs to be annotated.
            If we provide an Interface, this means that the function has been annotated partially,
            and we need to continue annotating the needed ones.
-
-        function_call_args : list[FunctionCallArgument], optional
-            The list of call arguments, needed only in the case of an inlined function.
         """
         if expr.get_direct_user_nodes(lambda u: isinstance(u, CodeBlock)):
-            errors.report("Functions can only be declared in modules or inside other functions.",
+            errors.report("Functions can only be declared in modules, classes or inside other functions.",
                     symbol=expr, severity='error')
 
         current_class = expr.get_direct_user_nodes(lambda u: isinstance(u, ClassDef))
@@ -4846,10 +4853,6 @@ class SemanticParser(BasicParser):
         is_elemental       = expr.is_elemental
         is_private         = expr.is_private
         is_inline          = expr.is_inline
-
-        if function_call_args is not None:
-            assert is_inline
-            found_func = False
 
         not_used = [d for d in decorators if d not in (*def_decorators.__all__, 'property', 'overload')]
         if len(not_used) >= 1:
@@ -4894,11 +4897,7 @@ class SemanticParser(BasicParser):
         interface_name = name
         interface_counter = 0
         is_interface = len(argument_combinations) > 1 or 'overload' in decorators
-        annotated_args = [] # collect annotated arguments to check for argument incompatibility errors
         for interface_idx, (arguments, type_var_idx) in enumerate(zip(argument_combinations, type_var_indices)):
-            if function_call_args is not None and found_func:
-                break
-
             if is_interface:
                 name, _ = self.scope.get_new_incremented_symbol(interface_name, interface_idx)
 
@@ -4915,20 +4914,6 @@ class SemanticParser(BasicParser):
                 self.scope.insert_symbolic_alias(n, dtype)
 
             arg_dict  = {a.name:a.var for a in arguments}
-            annotated_args.append(arguments)
-
-            if function_call_args is not None:
-                is_compatible = self._check_argument_compatibility(function_call_args, arguments, expr, is_elemental, raise_error=False)
-                if not is_compatible:
-                    self.exit_function_scope()
-                    # remove the new created scope and the function name
-                    self.scope.sons_scopes.pop(expr.name)
-                    if is_interface:
-                        self.scope.remove_symbol(name)
-                    continue
-                #In the case of an Interface we set found_func to True so that we don't continue
-                #searching for the other functions
-                found_func = True
 
             for a in arguments:
                 a_var = a.var
@@ -5081,12 +5066,6 @@ class SemanticParser(BasicParser):
             if expr.python_ast:
                 func.set_current_ast(expr.python_ast)
 
-        if function_call_args is not None and len(new_semantic_funcs) == 0:
-            for args in annotated_args[:-1]:
-                #raise errors if we do not find any compatible function def
-                self._check_argument_compatibility(function_call_args, args, expr, is_elemental, error_type='error')
-            self._check_argument_compatibility(function_call_args, annotated_args[-1], expr, is_elemental, error_type='fatal')
-
         if existing_semantic_funcs:
             new_semantic_funcs = existing_semantic_funcs + new_semantic_funcs
 
@@ -5105,6 +5084,173 @@ class SemanticParser(BasicParser):
             self.insert_function(new_semantic_funcs, insertion_scope)
 
         return EmptyNode()
+
+    def _visit_InlineFunctionDef(self, expr, function_call_args, function_call):
+        """
+        Visit an inline function definition to add the code to the calling scope.
+
+        Visit an inline function definition to add the code to the calling scope.
+        The code is inlined at this stage.
+
+        Parameters
+        ----------
+        expr : InlineFunctionDef
+            The inline function definition being called.
+        function_call_args : list[FunctionDefArgument]
+            The semantic arguments passed to the function.
+        function_call : FunctionCall
+            The syntactic function call being expanded to a function definition.
+        """
+        assign = function_call.get_direct_user_nodes(lambda a: isinstance(a, Assign) and not isinstance(a, AugAssign))
+        self._current_function.append(expr)
+        if assign:
+            lhs = assign[-1].lhs
+        else:
+            lhs = self.scope.get_new_name()
+        # Build the syntactic body
+        replace_map = {}
+
+        pyccel_stage.set_stage('syntactic')
+        global_scope_import_targets = {}
+        if expr.is_imported:
+            mod_name = expr.get_direct_user_nodes(lambda m: isinstance(m, Module))[0].name
+            mod = self.d_parsers[mod_name].semantic_parser.ast
+
+            global_symbols = set(expr.body.get_attribute_nodes(PyccelSymbol))
+            global_symbols.difference_update(expr.scope.local_used_symbols)
+
+            for v in global_symbols:
+                import_mod_name = mod_name
+                if mod.scope.find(v):
+                    imported_obj = None
+                    for import_type in mod.scope.imports.values():
+                        if v in import_type:
+                            imported_obj = import_type[v]
+                            break
+                    if imported_obj:
+                        import_mod_name = imported_obj.get_direct_user_nodes(lambda m: isinstance(m, Module))[0].name
+                    if self.scope.symbol_in_use(v):
+                        new_v = self.scope.get_new_name(self.scope.get_expected_name(v))
+                        replace_map[v] = new_v
+                        global_scope_import_targets.setdefault(import_mod_name, []).append(AsName(v, new_v))
+                    else:
+                        global_scope_import_targets.setdefault(import_mod_name, []).append(v)
+
+        # Swap in the function call arguments to replace the variables representing
+        # the arguments of the inlined function
+        res_vars = ()
+        if expr.results:
+            # Swap in the result of the function to replace the variable representing
+            # the result of the inlined function
+            res_var = expr.results.var
+            if isinstance(res_var, AnnotatedPyccelSymbol):
+                res_var = res_var.name
+            if isinstance(lhs, PyccelSymbol):
+                replace_map[res_var] = lhs
+                res_vars = (res_var,)
+
+        func_args = [a.var for a in expr.arguments]
+        func_args = [a.name if isinstance(a, AnnotatedPyccelSymbol) else a for a in func_args]
+
+        # Ensure local variables will be recognised and use a name that is not already in use
+        for v in expr.scope.local_used_symbols:
+            if v != expr.name and v not in res_vars:
+                if self.scope.symbol_in_use(v):
+                    new_v = self.scope.get_new_name(self.scope.get_expected_name(v))
+                    replace_map[v] = new_v
+                else:
+                    self.scope.insert_symbol(v)
+
+        # Map local call arguments to function arguments
+        positional_call_args = [a.value for a in function_call_args if not a.has_keyword]
+        for func_a, call_a in zip(func_args, positional_call_args):
+            if isinstance(call_a, Variable) and func_a == self.scope.get_expected_name(call_a.name):
+                # If call argument is a variable with the same name as the target function
+                # argument then there is no need to rename
+                new_func_a = replace_map.pop(func_a)
+                self.scope.remove_symbol(new_func_a)
+            else:
+                # Otherwise the symbol used for the function arguments should be mapped
+                # to the call argument
+                func_a_name = replace_map.get(func_a, func_a)
+                self.scope.variables[func_a_name] = call_a
+                self.scope.local_used_symbols[func_a_name] = func_a_name
+
+        # Map local keyword call arguments to function arguments
+        nargs = len(positional_call_args)
+        kw_call_args = {a.keyword: a.value for a in function_call_args[nargs:]}
+        for func_a, func_a_name in zip(expr.arguments[nargs:], func_args[nargs:]):
+            call_a = kw_call_args.get(func_a_name, getattr(func_a.default_call_arg, 'value', func_a.default_call_arg))
+            if isinstance(call_a, Variable) and func_a_name == self.scope.get_expected_name(call_a.name):
+                # If call argument is a variable with the same name as the target function
+                # argument then there is no need to rename
+                new_func_a = replace_map.pop(func_a_name)
+                self.scope.remove_symbol(new_func_a)
+            else:
+                # Otherwise the symbol used for the function arguments should be mapped
+                # to the call argument
+                used_func_a_name = replace_map.get(func_a_name, func_a_name)
+                self.scope.variables[used_func_a_name] = call_a
+                self.scope.local_used_symbols[func_a_name] = func_a_name
+
+        to_replace = list(replace_map.keys())
+        local_var = list(replace_map.values())
+        # Replace local syntactic variables from the inline functions with the syntactic
+        # variables defined above which are sure to not cause name collisions
+        expr.substitute(to_replace, local_var, invalidate = False)
+
+        # Replace return expressions with an assign to the results
+        returns = expr.body.get_attribute_nodes(Return)
+        replace_return = [Assign(lhs, r.expr, python_ast = r.python_ast) \
+                          if not isinstance(r.expr, PyccelSymbol) or not isinstance(lhs, PyccelSymbol) \
+                          else EmptyNode() for r in returns]
+        expr.body.substitute(returns, replace_return, invalidate = False)
+
+        imports = list(expr.imports)
+        imports.extend(Import(m_name, targets) for m_name, targets in global_scope_import_targets.items())
+        pyccel_stage.set_stage('semantic')
+
+        import_init_calls = [self._visit(i) for i in imports]
+
+        if expr.functions:
+            errors.report("Functions in inline functions are not supported",
+                    severity='error', symbol=expr)
+
+        # Visit the body as though it appeared directly in the code
+        body = self._visit(expr.body)
+        body.insert2body(*import_init_calls, back=False)
+
+        self._current_function.pop()
+
+        pyccel_stage.set_stage('syntactic')
+        # Put back the returns to create custom Assign nodes on the next visit
+        expr.body.substitute(replace_return, returns)
+
+        # Remove the symbol maps added to handle the function arguments
+        # These are found in self.scope.variables but do not represent variables
+        # that need to be declared.
+        for func_a, call_a in zip(func_args, positional_call_args):
+            func_a_name = replace_map.get(func_a, func_a)
+            if not isinstance(call_a, Variable) or func_a_name != call_a.name:
+                self.scope.remove_variable(call_a, func_a_name)
+
+        for func_a, func_a_name in zip(expr.arguments[nargs:], func_args[nargs:]):
+            if func_a_name in kw_call_args:
+                used_func_a_name = replace_map.get(func_a_name, func_a_name)
+                call_a = kw_call_args[func_a_name]
+                if not isinstance(call_a, Variable) or used_func_a_name != call_a.name:
+                    self.scope.remove_variable(call_a, used_func_a_name)
+
+        # Swap the arguments back to the original version to preserve the syntactic
+        # inline function definition.
+        expr.substitute(local_var, to_replace)
+        pyccel_stage.set_stage('semantic')
+
+        if assign:
+            return body
+        else:
+            self._additional_exprs[-1].append(body)
+            return self._visit(lhs)
 
     def _visit_PythonPrint(self, expr):
         args = [self._visit(i) for i in expr.expr]
@@ -5214,15 +5360,11 @@ class SemanticParser(BasicParser):
             var1, var2 = var2, var1
 
         if isinstance(var2, Nil):
-            if not isinstance(var1, Variable):
+            if not isinstance(var1, Variable) or not var1.is_optional:
                 if IsClass == PyccelIsNot:
                     return LiteralTrue()
                 elif IsClass == PyccelIs:
                     return LiteralFalse()
-            elif not var1.is_optional:
-                errors.report(PYCCEL_RESTRICTION_OPTIONAL_NONE,
-                        bounding_box=(self.current_ast_node.lineno, self.current_ast_node.col_offset),
-                        severity='error')
             return IsClass(var1, expr.rhs)
 
         if (var1.dtype != var2.dtype):
@@ -5373,6 +5515,7 @@ class SemanticParser(BasicParser):
 
                 if new_name != old_name:
                     import_init = import_init.clone(new_name)
+                    container['functions'][old_name] = import_init
 
                 result  = import_init()