+++ CHANGELOG.md
@@ -179,6 +179,7 @@ All notable changes to this project will be documented in this file.
 -   #2216 : Ensure compilation dependencies added by Pyccel are indicated for compilation of files which import the module.
 -   #2261 : Add LLVM compilers.
 -   #2344 : Allow language to be set using a capitalised name (Fortran, C, Python).
+-   #2322 : Fix inline functions calling inline functions with their own local variables.
 
 ### Changed
 
@@ -198,6 +199,8 @@ All notable changes to this project will be documented in this file.
 -   #2302 : Print the deallocation in a 1 line if statement.
 -   #2125 : Add information about received data type to type errors when calling a function with the wrong type.
 -   #297 : Parse generated `.pyi` files instead of `.py` files when importing to speed up translation.
+-   #2330 : Inline functions in the semantic stage.
+-   #2322 : Stop raising an error when checking if non-optional variable is `None`.
 -   \[INTERNALS\] `FunctionDef` is annotated when it is called, or at the end of the `CodeBlock` if it is never called.
 -   \[INTERNALS\] `InlinedFunctionDef` is only annotated if it is called.
 -   \[INTERNALS\] Build `utilities.metaclasses.ArgumentSingleton` on the fly to ensure correct docstrings.